+++ CHANGELOG.md
@@ -4,6 +4,11 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+## [4.2.0-52]
+### Added
+- GH-#12 Color chase buttons for easier finding
+- GH-#11 Reorder chase and redesigned the popup reorder for cues
+
 ## [4.1.0-51]
 ### Added
 - GH-#10 Add an import button in case your Android can't auto open the .AuroraDMX file